Update method to change Subdomain to root via Graph
Several customers opened support cases with Microsoft due to failure in converting child domains to root because they were not aware that the V1.0 of the graph is unable to do this (returns bad request). Suggesting the POST query to be edited to include "/beta/" in order to force the Beta version.

Leave the suggestion to perhaps update the other queries?